+++ TesteiOSSantander/TesteiOSSantander/Base.lproj/Main.storyboard
@@ -76,9 +76,9 @@
                     </view>
                     <navigationItem key="navigationItem" id="E2h-E1-lII"/>
                     <connections>
-                        <outlet property="loginLabel" destination="aJ7-zZ-D5q" id="Rbn-fy-TbM"/>
-                        <outlet property="passwordTextFild" destination="sCi-m5-A83" id="6jE-PG-hMt"/>
-                        <outlet property="userTextField" destination="SlP-hn-X7T" id="lXI-zA-6UP"/>
+                        <outlet property="loginLabel" destination="aJ7-zZ-D5q" id="6c3-vb-N5z"/>
+                        <outlet property="passwordTextFild" destination="sCi-m5-A83" id="qR2-7Y-Emd"/>
+                        <outlet property="userTextField" destination="SlP-hn-X7T" id="P6k-jP-Mgg"/>
                     </connections>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="dkx-z0-nzr" sceneMemberID="firstResponder"/>
