+++ TesteiOSSantander/TesteiOSSantanderUITests/TesteiOSSantanderUITests.swift
@@ -8,8 +8,8 @@
 
 import XCTest
 
-class TesteiOSSantanderUITests: XCTestCase {
 
+class TesteiOSSantanderUITests: XCTestCase {
     override func setUp() {
         // Put setup code here. This method is called before the invocation of each test method in the class.
